add: StringExtensions
